Refactoring. Some bugs fixed.
