+++ javascript_nodejs/src/bot.js
@@ -7,8 +7,28 @@
 // Import required Bot Builder
 const { ActivityHandler } = require('botbuilder');
 
+// Import required cognitive services
+const { LuisRecognizer, QnAMaker } = require('botbuilder-ai');
+
+// Import other required packages
+const { MainDialog } = require('./dialogs/main');
+const localizer = require('i18n');
+const appsettings = require('./appsettings.json');
+
 // State Accessor Properties
 const DIALOG_STATE_PROPERTY = 'dialogState';
+
+// Name of the QnA Maker service in the .bot file without the locale key.
+const QNA_CONFIGURATION = 'QNA-';
+
+// LUIS service type entry as defined in the .bot file without the locale key.
+const LUIS_CONFIGURATION = 'LUIS-';
+
+// CONSTS used in QnA Maker query.
+const QNA_MAKER_OPTIONS = {
+    scoreThreshold: 0.5,
+    top: 1
+};
 
 /**
  * Demonstrates the following concepts:
@@ -28,21 +48,52 @@ class CorePlusBot extends ActivityHandler {
      *
      * @param {ConversationState} conversationState property accessor
      * @param {UserState} userState property accessor
-     * @param {Dialog} dialog main dialog
      */
-    constructor(conversationState, userState, dialog) {
+    constructor(conversationState, userState) {
         super();
 
-        if (!conversationState) throw new Error('Missing parameter.  conversationState is required');
-        if (!userState) throw new Error('Missing parameter.  userState is required');
-        if (!dialog) throw new Error('Missing parameter. dialog is required');
+        if (!conversationState) throw new Error('Missing parameter. conversationState is required');
+        if (!userState) throw new Error('Missing parameter. userState is required');
 
         this.conversationState = conversationState;
         this.userState = userState;
-        this.dialog = dialog;
-        this.dialogState = this.conversationState.createProperty('DialogState');
 
-        this.rootDialogId = dialog.id;
+        const luisRecognizers = {};
+        const qnaRecognizers = {};
+        const availableLocales = localizer.getLocales();
+
+        // Add LUIS and QnAMaker recognizers for each locale
+        availableLocales.forEach((locale) => {
+            // Add the LUIS recognizer.
+            let luisConfig = appsettings[LUIS_CONFIGURATION + locale];
+
+            if (!luisConfig || !luisConfig.appId) {
+                throw new Error(`Missing LUIS configuration for locale "${ locale }" in appsettings.json file.\n`);
+            }
+
+            luisRecognizers[locale] = new LuisRecognizer({
+                applicationId: luisConfig.appId,
+                // CAUTION: Its better to assign and use a subscription key instead of authoring key here.
+                endpointKey: luisConfig.subscriptionKey,
+                endpoint: luisConfig.endpoint
+            }, undefined, true);
+
+            // Add the QnA recognizer.
+            let qnaConfig = appsettings[QNA_CONFIGURATION + locale];
+
+            if (!qnaConfig || !qnaConfig.kbId) {
+                throw new Error(`Missing QnA Maker configuration for locale "${ locale }" in appsettings.json file.\n`);
+            }
+
+            qnaRecognizers[locale] = new QnAMaker({
+                knowledgeBaseId: qnaConfig.kbId,
+                endpointKey: qnaConfig.endpointKey,
+                host: qnaConfig.hostname
+            }, QNA_MAKER_OPTIONS);
+        });
+
+        // Create the main dialog.
+        this.dialog = new MainDialog(luisRecognizers, qnaRecognizers, userState);
 
         // Create the property accessors for conversation state
         this.dialogState = conversationState.createProperty(DIALOG_STATE_PROPERTY);
